==> 1_County_DataTransformation.ipynb <==
# Import Dependencies
import pandas as pd
from pathlib import Path

national_county2024_csv = Path("./source/national_county2024.csv")

# Read the file, use "|" as a separator, preserve all column data types as text
source_df = pd.read_csv(national_county2024_csv, dtype=str)

source_df.head()

# Keep only required states
source_df = source_df[(source_df["STATEFP"] <= "56")]

# Remove unnecessary columns
source_df.drop(columns=["COUNTYNS"], inplace=True)

source_df.head()

# Create a set of "Unknown or Undefined" counties
unknown_county_df = source_df[["STATE", "STATEFP"]].drop_duplicates().copy()
unknown_county_df["COUNTYFP"] = 999
unknown_county_df["COUNTYNAME"] = "Unknown Or Undefined"

unknown_county_df.head()

# concatenate two DataFrames into a new one
county_df = pd.concat([source_df, unknown_county_df], ignore_index=True)

county_df.head()

county_df["county_fips"] = pd.to_numeric(county_df["STATEFP"].astype("str") + county_df["COUNTYFP"].astype("str"))

# Sort the DataFrame
county_df.sort_values(by=["county_fips"], inplace=True)

county_df.head()

# Remove unnecessary columns
county_df.drop(columns=["STATEFP", "COUNTYFP"], inplace=True)


# Rename columns
county_df.rename(columns={"STATE": "state_code",
                          "COUNTYNAME": "county_name"}, inplace=True)

# Reorder columns
county_df = county_df[["county_fips", "state_code", "county_name"]]

county_df.head(10)

# Save to file
county_df.to_csv(Path("./source/county.csv"), index=False, header=True)
