==> 2_BEA_DataTransformation.ipynb <==
# Import Dependencies
import pandas as pd
from pathlib import Path

# BEA Personal Consumption

bea_personal_consumption_csv = Path("./source/BEA/Personal_consumption_by_state_2020_2022.csv")

# Skip the first 3 lines
source_df = pd.read_csv(bea_personal_consumption_csv, header=3)

# Remove rows with NaN, since they are related to desciptions
source_df.dropna(inplace=True)

source_df.head(25)

# Keep only required LineCodes, related to main groups:
# 1 - Total
# 3 - Durable goods
# 8 - Nondurable goods
# 14 - Services
# 22 - NPISHs

source_df = source_df[(source_df["LineCode"].isin([1, 3, 8, 14, 22]))]

# Remove unnecessary columns
source_df.drop(columns=["GeoFips", "Description"], inplace=True)

source_df.head(10)

# Unpivot the table to convert year columns (2020, 2021, 2022) into rows
melted_df = pd.melt(source_df, id_vars=["GeoName", "LineCode"], var_name="Year", value_name="Value")

melted_df.head(10)

# Pivot the table to make each unique LineCode a column
bea_personal_consumption_df = melted_df.pivot_table(index=["GeoName", "Year"], columns="LineCode", values="Value", aggfunc='first').reset_index()

bea_personal_consumption_df.head(10)

bea_personal_consumption_df.columns

# Rename columns
bea_personal_consumption_df.rename(columns={"GeoName": "state_name",
                                            "Year": "year",
                                            1.0: "bea_pce_total",
                                            3.0: "bea_pce_durable_goods",
                                            8.0: "bea_pce_nondurable_goods",
                                            14.0: "bea_pce_services",
                                            22.0: "bea_pce_npish"}, inplace=True)

bea_personal_consumption_df.head(10)

# Save to file
bea_personal_consumption_df.to_csv(Path("./source/bea_personal_consumption.csv"), index=False, header=True)

# BEA Income and Population

bea_income_population_csv = Path("./source/BEA/Income_population_by_county_2020_2022.csv")

# Skip the first 3 lines
source_df = pd.read_csv(bea_income_population_csv, header=3)

# Remove rows with NaN, since they are related to desciptions
source_df.dropna(inplace=True)

source_df.head()

# Remove unnecessary columns
source_df.drop(columns=["GeoName", "Description"], inplace=True)

# Update data type
source_df["GeoFips"] = source_df["GeoFips"].astype(int)

source_df.head()

# Unpivot the table to convert year columns (2020, 2021, 2022) into rows
melted_df = pd.melt(source_df, id_vars=["GeoFips", "LineCode"], var_name="Year", value_name="Value")

melted_df.head()

# Pivot the table to make each unique LineCode a column
bea_income_population_df = melted_df.pivot_table(index=["GeoFips", "Year"], columns="LineCode", values="Value", aggfunc='first').reset_index()

bea_income_population_df.head(10)

# Rename columns
bea_income_population_df.rename(columns={"GeoFips": "county_fips",
                                         "Year": "year",
                                         10.0: "bea_total_income",
                                         20.0: "population"}, inplace=True)

# Filter out all records with population = (NA), use ~ as the NOT() condition
bea_income_population_df = bea_income_population_df[~(bea_income_population_df["population"] == "(NA)")]

# Calculate total income by multiplying by 1000
bea_income_population_df["bea_total_income"] = bea_income_population_df["bea_total_income"].astype("Int64") * 1000

bea_income_population_df.head()

# Save to file
bea_income_population_df.to_csv(Path("./source/bea_income_population.csv"), index=False, header=True)
