==> 4_county_metric_DataTransformation.ipynb <==
# Import Dependencies
import pandas as pd
from pathlib import Path

# Read BLS Employment
bls_employment_csv = Path("./source/bls_employment.csv")

bls_employment_df = pd.read_csv(bls_employment_csv)

bls_employment_df.head()

# Read BEA Income Population
bea_income_population_csv = Path("./source/bea_income_population.csv")

bea_income_population_df = pd.read_csv(bea_income_population_csv)

bea_income_population_df.head()

# Merge DataFrames
county_metric_df = pd.merge(bea_income_population_df, bls_employment_df)

county_metric_df.head()

# Save to file
county_metric_df.to_csv(Path("./source/county_metric.csv"), index=False, header=True)
