==> 6_LoadToSQLite.ipynb <==
# Import Dependencies
import pandas as pd
from pathlib import Path
from sqlalchemy import create_engine, text

# Read source data

# Read state
state_csv = Path("./source/state.csv")

state_df = pd.read_csv(state_csv)

state_df.head()

# Read county
county_csv = Path("./source/county.csv")

county_df = pd.read_csv(county_csv)

county_df.head()

# Read industry
industry_csv = Path("./source/industry.csv")

industry_df = pd.read_csv(industry_csv)

industry_df.head()

# Read county_metric
county_metric_csv = Path("./source/county_metric.csv")

county_metric_df = pd.read_csv(county_metric_csv)

county_metric_df.head()

# Read county_industry_metric
county_industry_metric_csv = Path("./source/county_industry_metric.csv")

county_industry_metric_df = pd.read_csv(county_industry_metric_csv)

county_industry_metric_df.head()

# Upload data to the database

# Connect to database
database_path = Path("data.sqlite")

engine = create_engine(f"sqlite:///{database_path}")

# state
state_df.to_sql("state", engine, if_exists='replace', index=False, method='multi')

# county
county_df.to_sql("county", engine, if_exists='replace', index=False, method='multi')

# industry
industry_df.to_sql("industry", engine, if_exists='replace', index=False, method='multi')

# county_metric
county_metric_df.to_sql("county_metric", engine, if_exists='replace', index=False, method='multi')

# county_industry_metric - method='multi' produces an errror "OperationalError: (sqlite3.OperationalError) too many SQL variables"
county_industry_metric_df.to_sql("county_industry_metric", engine, if_exists='replace', index=False)
